fix(marketing): make the marquee pause real and correct the RSS and consent gating
The pause affordance never worked, in either mode. Tailwind was not generating
the group-hover/group-focus-within arbitrary-property utilities, so no
animation-play-state rule reached the page at all, and the animation shorthand
carried running. The rule now lives beside the keyframes, scoped to
[data-marquee], and the strip is focusable so a keyboard can reach it. Verified:
running, focus, paused, blur, running.

RSS: <author> requires an email address in RSS 2.0, so a display name there is
invalid; it becomes dc:creator with the namespace declared. URLs are escaped
like the titles already were, so a slug or base URL carrying an ampersand cannot
break the document.

The consent line sat inside the hasAlternatives branch, so a magic-link-only
fork that had set site.legal still saw nothing; it is now a sibling.

Client-side email normalization was redundant: the waitlist router already trims
in its schema and lowercases on insert. The comment claiming otherwise is gone,
and only the trim used for the echoed confirmation remains.

Splits code-window.tsx into three single-export files, the only shape the naming
rule allows. Adds the feed to the roadmap doc, and scopes the footer rule to
surfaces that actually have footer content.
